missing modules imported, the additional content in the test_getter.py moved to the test_config.py
